docs: moves mocks into tests directory
